Rename the iterators submodule to shapetools
The name doesn't really matter because people should only be using the
top-level namespace for public API functions. But shapetools makes more sense
given the functions that are actually there.
